change return type from Int to Long

+++ data/src/main/java/io/github/golok56/data/db/CategoryDao.kt
@@ -10,7 +10,7 @@ interface CategoryDao {
     fun getAll(): Single<List<CategoryData>>
 
     @Insert
-    fun insert(category: CategoryData): Int
+    fun insert(category: CategoryData): Long
 
     @Delete
     fun delete(category: CategoryData): Int
